merkle: Move rehashing algorithm in (#2186)
This change moves the rehasher algorithm to merkle package, and refactors it
to be a single function.
