+++ merkle/log_proofs.go
@@ -15,6 +15,7 @@
 package merkle
 
 import (
+	"errors"
 	"fmt"
 	"math/bits"
 
@@ -46,6 +47,8 @@ func checkSnapshot(ssDesc string, ss, treeSize int64) error {
 // inclusion proof for a specified leaf and tree size. The snapshot parameter
 // is the tree size being queried for, treeSize is the actual size of the tree
 // at the revision we are using to fetch nodes (this can be > snapshot).
+//
+// Use Rehash function to compose the proof after the node hashes are fetched.
 func CalcInclusionProofNodeAddresses(snapshot, index, treeSize int64) ([]NodeFetch, error) {
 	if err := checkSnapshot("snapshot", snapshot, treeSize); err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "invalid parameter for inclusion proof: %v", err)
@@ -71,6 +74,8 @@ func CalcInclusionProofNodeAddresses(snapshot, index, treeSize int64) ([]NodeFet
 // to valid tree heads. All returned NodeIDs are tree coordinates within the
 // new tree. It is assumed that they will be fetched from storage at a revision
 // corresponding to the STH associated with the treeSize parameter.
+//
+// Use Rehash function to compose the proof after the node hashes are fetched.
 func CalcConsistencyProofNodeAddresses(snapshot1, snapshot2, treeSize int64) ([]NodeFetch, error) {
 	if err := checkSnapshot("snapshot1", snapshot1, treeSize); err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "invalid parameter for consistency proof: %v", err)
@@ -171,6 +176,33 @@ func proofNodes(index uint64, level uint, size uint64, rehash bool) []NodeFetch 
 	}
 
 	return proof
+}
+
+// Rehash computes the proof based on the slice of NodeFetch structs, and the
+// corresponding hashes of these nodes. The slices must be of the same length.
+// The hc parameter computes node's hash based on hashes of its children.
+//
+// Warning: The passed-in slice of hashes can be modified in-place.
+func Rehash(h [][]byte, nf []NodeFetch, hc func(left, right []byte) []byte) ([][]byte, error) {
+	if len(h) != len(nf) {
+		return nil, errors.New("slice lengths mismatch")
+	}
+	cursor := 0
+	// Scan the list of node hashes, and store the rehashed list in-place.
+	// Invariant: cursor <= i, and h[:cursor] contains all the hashes of the
+	// rehashed list after scanning h up to index i-1.
+	for i, ln := 0, len(h); i < ln; i, cursor = i+1, cursor+1 {
+		hash := h[i]
+		if nf[i].Rehash {
+			// Scan the block of node hashes that need rehashing.
+			for i++; i < len(nf) && nf[i].Rehash; i++ {
+				hash = hc(h[i], hash)
+			}
+			i--
+		}
+		h[cursor] = hash
+	}
+	return h[:cursor], nil
 }
 
 func reverse(ids []compact.NodeID) []compact.NodeID {
