+++ merkle/log_proofs_test.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/trillian/merkle/compact"
+	"github.com/google/trillian/merkle/rfc6962"
 )
 
 // TestCalcInclusionProofNodeAddresses contains inclusion proof tests. For
@@ -290,6 +291,72 @@ func TestConsistencySucceedsUpToTreeSize(t *testing.T) {
 				t.Errorf("CalcConsistencyProofNodeAddresses(%d, %d) = %v", s1, s2, err)
 			}
 		}
+	}
+}
+
+func TestRehasher(t *testing.T) {
+	th := rfc6962.DefaultHasher
+	h := [][]byte{
+		th.HashLeaf([]byte("Hash 1")),
+		th.HashLeaf([]byte("Hash 2")),
+		th.HashLeaf([]byte("Hash 3")),
+		th.HashLeaf([]byte("Hash 4")),
+		th.HashLeaf([]byte("Hash 5")),
+	}
+
+	for _, tc := range []struct {
+		desc   string
+		hashes [][]byte
+		rehash []bool
+		want   [][]byte
+	}{
+		{
+			desc:   "no rehash",
+			hashes: h[:3],
+			rehash: []bool{false, false, false},
+			want:   h[:3],
+		},
+		{
+			desc:   "single rehash",
+			hashes: h[:5],
+			rehash: []bool{false, true, true, false, false},
+			want:   [][]byte{h[0], th.HashChildren(h[2], h[1]), h[3], h[4]},
+		},
+		{
+			desc:   "single rehash at end",
+			hashes: h[:3],
+			rehash: []bool{false, true, true},
+			want:   [][]byte{h[0], th.HashChildren(h[2], h[1])},
+		},
+		{
+			desc:   "single rehash multiple nodes",
+			hashes: h[:5],
+			rehash: []bool{false, true, true, true, false},
+			want:   [][]byte{h[0], th.HashChildren(h[3], th.HashChildren(h[2], h[1])), h[4]},
+		},
+		{
+			// TODO(pavelkalinnikov): This will never happen in our use-case. Design
+			// the type to not allow multi-rehash by design.
+			desc:   "multiple rehash",
+			hashes: h[:5],
+			rehash: []bool{true, true, false, true, true},
+			want:   [][]byte{th.HashChildren(h[1], h[0]), h[2], th.HashChildren(h[4], h[3])},
+		},
+	} {
+		t.Run(tc.desc, func(t *testing.T) {
+			nf := make([]NodeFetch, len(tc.rehash))
+			for i, r := range tc.rehash {
+				nf[i].Rehash = r
+			}
+			h := append([][]byte{}, tc.hashes...)
+			got, err := Rehash(h, nf, th.HashChildren)
+			if err != nil {
+				t.Errorf("Rehash: %v", err)
+			}
+			if want := tc.want; !cmp.Equal(got, want) {
+				t.Errorf("proofs mismatch:\ngot: %x\nwant: %x", got, want)
+			}
+		})
 	}
 }
